Removed user key for EmailJS to .env, form now resets, added alert to confirm sent
