+++ .env
@@ -1,1 +1,3 @@
-PUBLIC_URL=https://bruggineer.github.io/PortfolioUpdated/public
+REACT_APP_USER="user_q7Y4TwUydLQKj2h4J96gv"
+PUBLIC_URL="https://bruggineer.github.io/PortfolioUpdated/public"
+
