+++ public/index.html
@@ -26,14 +26,6 @@
     -->
   <!-- Bootstrap-->
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
-  <!-- EmailJS for form
-  <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/emailjs-com@2/dist/email.min.js">
-  </script>
-  <script type="text/javascript">
-    (function () {
-      emailjs.init("user_q7Y4TwUydLQKj2h4J96gv");
-    })();
-  </script> -->
     <title>React Portfolio</title>
 </head >
