+++ src/index.js
@@ -5,6 +5,4 @@ import App from './App';
 
 console.log('public url: ', process.env.PUBLIC_URL)
 
-
-
 ReactDOM.render(<App />, document.getElementById("root"));
